% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_doremitra_datasets.R
\name{fetch_all_DoReMiTra}
\alias{fetch_all_DoReMiTra}
\title{Title}
\usage{
fetch_all_DoReMiTra(verbose = TRUE)
}
\arguments{
\item{verbose}{Logical. Whether to print progress messages. Default is TRUE.}
}
\value{
A list storing all the SE objects
}
\description{
Title
}
\examples{
fetch_all_DoReMiTra()

}
