% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_radiation_data.R
\name{get_DoReMiTra_data}
\alias{get_DoReMiTra_data}
\title{Title}
\usage{
get_DoReMiTra_data(dataset_name)
}
\arguments{
\item{dataset_name}{Character. The exact name of the dataset (e.g. "SE_Amundson_2008_ExVivo_GSE8917_GPL1708").}
}
\value{
A \code{SummarizedExperiment} object fetched from ExperimentHub.
}
\description{
Title
}
\examples{
get_DoReMiTra_data("SE_Amundson_2008_ExVivo_GSE8917_GPL1708")

}
