% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_datasets.R
\name{list_datasets}
\alias{list_datasets}
\title{List Available DoReMiTra Datasets
Displays a summary of available datasets with key metadata fields.}
\usage{
list_datasets()
}
\value{
A data.frame with metadata for each dataset
}
\description{
List Available DoReMiTra Datasets
Displays a summary of available datasets with key metadata fields.
}
\examples{
list_datasets()


}
