% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_datasets.R
\name{list_DoReMiTra_datasets}
\alias{list_DoReMiTra_datasets}
\title{List Available DoReMiTra Datasets
Displays a summary of available datasets with key metadata fields.}
\usage{
list_DoReMiTra_datasets(show_all_fields = FALSE)
}
\arguments{
\item{show_all_fields}{Logical. if TRUE, it returns all the metadata information}
}
\value{
A data.frame with metadata for each dataset
}
\description{
List Available DoReMiTra Datasets
Displays a summary of available datasets with key metadata fields.
}
\examples{
list_DoReMiTra_datasets()


}
