% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_DoReMiTra_metadata_fields.R
\name{list_DoReMiTra_metadata_fields}
\alias{list_DoReMiTra_metadata_fields}
\title{List available metadata fields and their values in DoReMiTra}
\usage{
list_DoReMiTra_metadata_fields()
}
\value{
Invisibly returns a named list of metadata fields and their values.
}
\description{
This function displays the available metadata fields across all
DoReMiTra datasets along with the unique values observed for each field.
It facilitates intuitive dataset discovery and filtering.
}
\examples{
list_DoReMiTra_metadata_fields()

}
