% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_doremitra_datasets.R
\name{load_doremitra_datasets}
\alias{load_doremitra_datasets}
\title{Title}
\usage{
load_doremitra_datasets(verbose = TRUE)
}
\arguments{
\item{verbose}{}
}
\value{
all SummarizedExperiment objects and stores them in the global environment
}
\description{
Title
}
\examples{
load_doremitra_datasets()

}
