% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/se_summary.R
\name{se_summary}
\alias{se_summary}
\title{Summarize Metadata for a DoReMiTra Dataset}
\usage{
se_summary(dataset_name)
}
\arguments{
\item{dataset_name}{Character string specifying the name of the dataset}
}
\value{
A character string containing a essential information about the dataset metadata
}
\description{
Summarize Metadata for a DoReMiTra Dataset
}
\examples{
se_name()
se_summary()
}
