% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DoReMiTra-pkg.R
\docType{package}
\name{DoReMiTra-pkg}
\alias{DoReMiTra-package}
\alias{DoReMiTra-pkg}
\title{\verb{DoReMiTra:} radiation DOse REsponse Measured In TRAnscriptomics}
\description{
\code{DoReMiTra} is a Bioconductor package that provides curated and harmonized radiation transcriptomic datasets
from publicly available radiation studies. It enables easy access to
dose- and time-dependent gene expression data across various organisms,
radiation types, and experimental settings.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/AhmedSAHassan/DoReMiTra}
  \item Report bugs at \url{https://github.com/AhmedSAHassan/DoReMiTra/issues}
}

}
\author{
\strong{Maintainer}: Ahmed Salah \email{ahhassan@uni-mainz.de} (\href{https://orcid.org/0000-0001-7003-4466}{ORCID})

Authors:
\itemize{
  \item Federico Marini \email{marinif@uni-mainz.de} (\href{https://orcid.org/0000-0003-3252-7758}{ORCID})
}

}
\keyword{internal}
