% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_DoReMiTra_datasets.R
\name{list_DoReMiTra_datasets}
\alias{list_DoReMiTra_datasets}
\title{List all available DoReMiTra datasets with associated key metadata information}
\usage{
list_DoReMiTra_datasets(show_all_fields = FALSE)
}
\arguments{
\item{show_all_fields}{Logical. If TRUE, it returns all the metadata information}
}
\value{
A data.frame with metadata for each dataset
}
\description{
Returns a metadata dataframe of all datasets available in the DoReMiTra collection,
including details such as title, organism, radiation type, experimental setting,
and accession numbers. Can optionally display extended metadata fields.
}
\examples{
list_DoReMiTra_datasets()


}
